Add a contrived params example
Most of the time you wouldn't need to access routing params from
a component because the match object that includes them is
available as prop in route components. This example is just to
demonstrate the access of the match object via the redux state.
This is sometimes useful to have, e.g. in a tracking middleware.